% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ajive}
\alias{ajive}
\title{Angle based Joint and Individual Variation Explained}
\usage{
ajive(
  blocks,
  initial_signal_ranks,
  full = TRUE,
  n_wedin_samples = 1000,
  n_rand_dir_samples = 1000,
  joint_rank = NA
)
}
\arguments{
\item{blocks}{List. A list of the data matrices.}

\item{initial_signal_ranks}{Vector. The initial signal rank estimates.}

\item{full}{Boolean. Whether or not to store the full J, I, E matrices or just their SVDs (set to FALSE to save memory).}

\item{n_wedin_samples}{Integer. Number of wedin bound samples to draw for each data matrix.}

\item{n_rand_dir_samples}{Integer. Number of random direction bound samples to draw.}

\item{joint_rank}{Integer or NA. User specified joint_rank. If NA will be estimated from data.}
}
\value{
The JIVE decomposition.
}
\description{
Computes the JIVE decomposition.
}
