% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_data}
\alias{gene_data}
\title{Example gene expression data}
\format{
## `gene_data`
A data frame with 2000 rows (genes) and 100 columns (sample).
}
\usage{
gene_data
}
\description{
A subset of RNA-seq data PsychENCODE
}
\keyword{datasets}
