% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta}
\alias{meta}
\title{Metadata}
\format{
## `meta`
A data frame with 100 rows (sample) and 2 columns (disease status and gender).
}
\usage{
meta
}
\description{
Metadata with interested phenotype.
}
\keyword{datasets}
