% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prop_gene}
\alias{prop_gene}
\title{Deconvoluted proportion based on gene expression}
\format{
## `prop_gene`
A data frame of cell-type proportion estimates for each sample based on RNA-seq expression.
}
\usage{
prop_gene
}
\description{
Cell-type proportion estimates from gene expression
}
\keyword{datasets}
