% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_gene}
\alias{ref_gene}
\title{Signature matrix}
\format{
## `ref_gene`
A data frame with 4872 rows (genes) and 5 columns (cell types).
}
\usage{
ref_gene
}
\description{
A signature matrix which serves as a reference of known cellular signatures.
}
\keyword{datasets}
