% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{deconv}
\alias{deconv}
\title{Estimation of cellular composition in high-throughput data from heterogeneous tissues}
\usage{
deconv(se, source = "protein", method = "cibersort")
}
\arguments{
\item{se}{A `SummarizedExperiment` object with bulk protein expression data frame contained in `counts` slot, and
a "signature matrix" which serves as a reference of known cellular signatures contained as an element (`sig_matrix`) in `metadata` slot.}

\item{source}{A character string denotes which molecular profiles to be deconvoluted. The setting of `proteins` or `transcript` means single-source 
deconvolution with source-specific signature matrix, while 'cross' means proteome deconvolution based on pre-estimated transcriptome proportion.}

\item{method}{A character string denotes which deconvolution method to use. In the current version, only `CIBERSORT` or `nnls` is supported.}
}
\value{
A `SummarizedExperiment`. The cell-type proportion estimates for each sample will be stored as an element (`prop`) in `metadata` slot.
}
\description{
This function returns a `SummarizedExperiment` object including cell-type proportion estimates for each sample.
}
\details{
This is a function developed to implement cell-type proportion deconvolution using either `CIBERSORT` or `nnls`.
}
\examples{

se <- SummarizedExperiment(assays = list(protein = mcQTL::protein_data),
                           rowData = mcQTL::anno_protein)
metadata(se) <- list(sig_matrix = mcQTL::ref_protein)
se <- deconv(se, "cibersort")
head(se@metadata$prop)


}
